use a OCSController for the external API
