+++ appinfo/info.xml
@@ -9,6 +9,6 @@
     <namespace>Survey_Server</namespace>
     <category>other</category>
     <dependencies>
-        <nextcloud min-version="9" max-version="13"/>
+        <nextcloud min-version="9" max-version="14"/>
     </dependencies>
 </info>
